+++ change_log.md
@@ -2,6 +2,8 @@ SynDaver Symple Slicer v0.9.93
 ------------------------------
 
 - Added the ability to make lithophanes out of images.
+- Reduce size of extremely large objects to fit
+- Fixed objects not falling to the bed.
 
 SynDaver Symple Slicer v0.9.92
 ------------------------------